requirement.txt
fastapi
uvicorn
python-multipart
pydantic
pandas
numpy
pillow
pdf2image
pytesseract
opencv-python-headless
transformers
torch
joblib
conductor-python
accelerate
evaluate
nltk
spacy
pdfplumber
datasets
scikit-learn
